% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{Phrasal Verbs}
\format{A tibble with 1108 rows and two columns:
\describe{
   \item{PHRASES}{all phrasal verbs in English Club website}
   \item{MEANING}{meaning of those phrasal verbs}
}}
\source{
\url{https://www.englishclub.com/ref/Phrasal_Verbs/}
}
\usage{
phrases
}
\description{
Contains 1108 phrasal verbs from English Club as of 14th April 2018
}
\keyword{datasets}
